[ADD] 직원 생성시 휴가 정보 추가, vac info service, repository 추가

+++ src/main/java/com/konai/kurong/faketee/vacation/dto/VacInfoResponseDto.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class VacInfoResponseDto {
 
     private Long id;
-    private Long remain;
+    private Double remain;
     private EmployeeResponseDto employeeResponseDto;
     private VacGroupResponseDto vacGroupResponseDto;
